Starting RDT
Added function that converts a string to a list of binary characters using utf-8 encoding. Added a function to convert it back to a string. It doesn't work for more fancy characters yet because the decoding is actually utf-16, but it should be fine for our use case. Now, I have a foundation to start checksum with.